Added support for COMB files
The COMB files I downloaded have email addresses and passwords with : as the separators.

The function I've added will import the passwords.

The Add-MissingHashes function will then need to be called to create the missing hashes.

The other minor change was to move the SQL DB creation into its own function.

Screen out was also tweaked.